requirements.txt
appdirs==1.4.3
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.3.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
dotty-dict==1.2.1
entrypoints==0.3
flake8==3.7.8
flake8-colors==0.1.6
idna==2.8
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.730
mypy-extensions==0.4.2
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
pyparsing==2.4.2
pytest==5.2.1
PyYAML==5.1.2
requests==2.22.0
setuptools-scm==3.3.3
six==1.12.0
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
